Move Scheduler files to folder
